refactor: Bring platform handlers outside of atomizer module.
